Here are some libraries that we will use for our projects ->ADS7485 is for touch screen driver ->digitalWriteFast some improved functions for AVR for digital write ->GraphicsLib grafic library that contains some fonts too ->MI0283QT2 the library for the display, so we can control it ->mSD_Shield library for the MicroSD shield so we can add an memory card to store images or something similar (if necessary)
This are some of the library, probably in the future we will add some other libraries. To be able to use the you have to copy them in "libraries" forder from your Arduino instalation path and restart te IDE.

PS: it contains some examples too, so we will understand easier how is working.