Design Phase Request from moodle
This is the basic schematic of the project. Because in Fritzing you can't find the shield that we use, I've just added an Arduino Mega 2560 and a mSD card shield, somehow they are similar. Because we don't have any external circuit in this project is hard to make a schematic using Fritzing but I made some connections for power and for SPI communication between display and Arduino whitch is tha main thing. All the analog pins from Arduino will be the input of the osciloscope .